Move monitor helper methods into utils
